requirements.txt
yfinance
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
tensorflow==2.16.1
xgboost==2.0.3
optuna==3.5.0
vaderSentiment==3.3.2
feedparser==6.0.11
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.19.0
pyyaml==6.0.1
joblib==1.3.2
jupyter==1.0.0
streamlit
